Rename views.py to viewsets.py
